requirements.txt
flask
requests
cryptography
PyNaCl
